requirements.txt
flask==2.2.2
flask-cors==3.0.10  # Ensure this is the compatible version
groq==0.13.0
python-dotenv==1.0.1
requests==2.32.3
werkzeug==2.2.3  # Add this line to fix the compatibility issue


# Backend : htts://xyz.com
langchain==0.1.16
langchain-core
langchain-groq